change platform

+++ Library/STC32G_Delay.h
@@ -1,8 +1,13 @@
+/*---------------------------------------------------------------------*/
+/* --- Web: www.STCAI.com ---------------------------------------------*/
+/*---------------------------------------------------------------------*/
+
 #ifndef	__STC32G_DELAY_H
 #define	__STC32G_DELAY_H
 
 #include	"config.h"
 
 void delay_ms(unsigned int ms);
-void delay_us(u16 us);
+void delay_us(unsigned int us);
+
 #endif
